Move equally_weighted, improve Portfolio.value, remove Portfolio.amount
