+++ src/backlight/portfolio/strategy.py
@@ -1,11 +1,16 @@
 # Define how to create Portfolio
 from typing import List
-from backlight.portfolio.portfolio import Portfolio
+from backlight.portfolio.portfolio import (
+    Portfolio,
+    _get_forex_ratios,
+    construct_portfolio,
+)
 from backlight.positions.positions import Positions
 from backlight.trades.trades import Trades
 from backlight.signal.signal import Signal
 from backlight.datasource.marketdata import MarketData
 from backlight import strategies
+from backlight.asset.currency import Currency
 
 from joblib import Parallel, delayed
 
@@ -41,3 +46,46 @@ def generate_simple_trades(
     )
 
     return trades
+
+
+def equally_weighted_portfolio(
+    trades: List[Trades],
+    mkt: List[MarketData],
+    principal: float,
+    max_amount: float,
+    currency_unit: Currency = Currency.USD,
+) -> Portfolio:
+    """
+    Create a Portfolio from trades and mkt, given a principal which will be divided equally between the 
+    different currencies.
+    args :
+        - trades : a list of trades for each currencies
+        - mkt : the market datas for at least each trades currencies
+        - principal : the total amount allocated to the Portfolio
+        - max_amount : the max amount
+        - currency_unit : the unit type of the future Portfolio
+    """
+    nb_currencies = len(trades)
+
+    symbols = [t.symbol for t in trades]
+    symbols2mkt = {m.symbol: m for m in mkt}
+    symbols2tds = {t.symbol: t for t in trades}
+
+    principals = {}
+    #     max_amounts = {}
+    lts = {}
+    for trade in trades:
+        symbol = trade.symbol
+        starting_date = trade.index[0]
+        trade_currency = trade.currency_unit
+
+        ratio = 1
+        if trade_currency != currency_unit:
+            ratios = _get_forex_ratios(mkt, trade_currency, currency_unit)
+            ratio = ratios.iloc[ratios.index.get_loc(trade.index[0]) - 1]
+
+        count_symbol = symbols.count(symbol)
+        principals[symbol] = principal / (nb_currencies * ratio * count_symbol)
+        lts[symbol] = int(principal / (max_amount))
+
+    return construct_portfolio(trades, mkt, principals, lts, currency_unit)
