+++ tests/portfolio/test_portfolio.py
@@ -4,7 +4,8 @@ import numpy as np
 
 import backlight
 from backlight.portfolio.portfolio import construct_portfolio as module
-from backlight.portfolio.portfolio import _fusion_positions, equally_weighted_portfolio
+from backlight.portfolio.portfolio import _fusion_positions
+from backlight.portfolio.strategy import equally_weighted_portfolio
 import backlight.positions.positions
 from backlight.trades.trades import make_trades
 from backlight.asset.currency import Currency
